requirements.txt
pandas
streamlit
plotly
pyodbc
azure-identity